Refactor experimental fluent dependencies
It is better to keep the dsl test as they were and put the model
test in the lambda module
